updated controllers and added contactUs

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Course = require('../models/Course');
 const Section = require('../models/Section');
+const SubSection = require("../models/Subsection")
 
 
 exports.createSection = async (req , res)=>{
@@ -19,7 +20,7 @@ exports.createSection = async (req , res)=>{
         // create section
         const newSection = await Section.create({sectionName})
         // update course with section id
-        const updatedCourseDetails = await Course.findByIdAndUpdate(
+        const updatedCourse = await Course.findByIdAndUpdate(
                                                     courseId,
                                                     {
                                                         $push:{
@@ -40,7 +41,7 @@ exports.createSection = async (req , res)=>{
         return res.status(200).json({
             success:true,
             message:"Section created Successfully ",
-            updatedCourseDetails,
+            updatedCourse,
         })
 
     } catch (error) {
@@ -60,26 +61,27 @@ exports.updateSection = async (req , res) => {
     try {
         
         // data input
-        const {sectionName , courseId} = req.body;
-        // data validation
-        if(!sectionName || !courseId)
-        {
-            return res.status(400).json({
-                success:false,
-                message:"Missing Properties",
-
-            })
-        }
+        const { sectionName, sectionId, courseId } = req.body
         // update data
         const section = await Section.findByIdAndUpdate(
-                                                    courseId,
+                                                    sectionId,
                                                     {sectionName},
                                                     {new:true},
                                                 )
+        const course = await Course.findById(courseId)
+        .populate({
+          path: "courseContent",
+          populate: {
+            path: "subSection",
+          },
+        })
+        .exec()
+        console.log(course)
         // retrun response
         return res.status(200).json({
             success:true,
-            message:"Section updated Successfully ",
+            message: section,
+            data: course,
         })
 
     } catch (error) {
@@ -93,27 +95,48 @@ exports.updateSection = async (req , res) => {
 }
 
 // ****** Delete section *****
-exports.deleteSection = async (req , res) => {
-
+exports.deleteSection = async (req, res) => {
     try {
-        
-        // get ID -- assuming that we are sending id in params
-        const {sectionId} = req.params;
-        // use findByIdAndDelete
-        await Section.findByIdAndDelete(sectionId);
-        //TODO [testing]:do we need to delete the entry from course Schema
-        // return response
-        return res.status(200).json({
-            success:true,
-            message:"Section Deleted Successfully ",
+      const { sectionId, courseId } = req.body
+      await Course.findByIdAndUpdate(courseId, {
+        $pull: {
+          courseContent: sectionId,
+        },
+      })
+      const section = await Section.findById(sectionId)
+      console.log(sectionId, courseId)
+      if (!section) {
+        return res.status(404).json({
+          success: false,
+          message: "Section not found",
         })
-
+      }
+      // Delete the associated subsections
+      await SubSection.deleteMany({ _id: { $in: section.subSection } })
+  
+      await Section.findByIdAndDelete(sectionId)
+  
+      // find the updated course and return it
+      const course = await Course.findById(courseId)
+        .populate({
+          path: "courseContent",
+          populate: {
+            path: "subSection",
+          },
+        })
+        .exec()
+  
+      res.status(200).json({
+        success: true,
+        message: "Section deleted",
+        data: course,
+      })
     } catch (error) {
-        return res.status(500).json({
-            success:false,
-            message:"Unable To Delete Section , Please try again",
-            error:error.message,
-        })
+      console.error("Error deleting section:", error)
+      res.status(500).json({
+        success: false,
+        message: "Internal server error",
+        error: error.message,
+      })
     }
-
-}
+  }
